# swaggerDocumentation.yaml
openapi: 3.0.0
info:
  title: Ticketline API
  description: Documentation for the Ticketline API
  version: 0.0.1
servers:
  - url: localhost
    description: Can be used when spring application is started locally
paths:
...
  /readArticleStatus/{id}:
    put:
      tags:
        - articles
      summary: Sets the status to read or unread for an article for the current user
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/articleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: boolean
      responses:
        200:
          description: Successful set the "read"-status of an article for an user
        401:
          $ref: "#/components/responses/401-not-authenticated"
        500:
          $ref: "#/components/responses/500-internal-error"
...
  /seatingPlans:
    post:
      tags: 
        - seatingPlans
      summary: Creates a seating plan.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatingPlanWithoutId"
      responses:
        201:
          description: Successful creation of a seating plan
          headers: 
            Location:
              $ref: "#/components/headers/Location"
        401:
          $ref: "#/components/responses/401-not-authenticated"
        403:
          $ref: "#/components/responses/403-forbidden"
        422:
          description: Validation failed for an input
        500:
          $ref: "#/components/responses/500-internal-error"
    get:
      tags: 
        - seatingPlans
      summary: Gets all seating plans
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful retreival of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeatingPlan"
        401:
          $ref: "#/components/responses/401-not-authenticated"
        500:
          $ref: "#/components/responses/500-internal-error"
  /seatingPlans/{id}:
    get:
      tags: 
        - seatingPlans
      summary: Retreives information of the seating plan with the given ID.
      parameters:
        - $ref: "#/components/parameters/seatingPlanId"
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful retreival of a seating plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatingPlan"
        401:
          $ref: "#/components/responses/401-not-authenticated"
        404:
          description: The seating plan with the given ID was not found
        500:
          $ref: "#/components/responses/500-internal-error"
...
components:
  parameters:
...
    articleId:
      name: id
      in: path
      description: ID of the user that is retreived
      required: true
      schema:
        type: integer
...
    seatingPlanId:
      name: id
      in: path
      description: ID of the seating plan layout that is retreived
      required: true
      schema:
        type: integer
...
  responses:
    500-internal-error:
      description: Internal Server Error
    401-not-authenticated:
      description: The user is not logged in
    403-forbidden:
      description: The user needs administrative rights
...
  headers:
    Location:
      required: true
      description: The location of the created resource
      schema:
        type: string
        format: uri
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SeatingPlanWithoutId:
      type: object
      properties:
        name:
          type: string
        seatingPlanLayoutId:
          type: number
        locationId:
          type: integer
      required:
        - name
        - seatingPlanLayoutId
        - locationId
    SeatingPlan:
      type: object
      properties:
        seatingPlanId:
          type: integer
        name:
          type: string
        seatingPlanLayoutId:
          type: number
        locationId:
          type: integer
      required:
        - seatingPlanId
        - name
        - seatingPlanLayoutId
        - locationId
...